Fatmug/requirements.txt
asgiref==3.7.2
autopep8==2.0.4
black==23.11.0
click==8.1.7
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pycodestyle==2.11.1
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
